backend/requirements.txt
# FastAPI & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI/ML
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
redis==5.0.1

# WhatsApp/Notifications
twilio==8.13.0
pywhatkit==5.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Deployment
gunicorn==21.2.0
python-dateutil==2.8.2
holidays==0.34